add note on navigation and optionall copy the LICENSE file down

+++ bin/entrypoint.sh
@@ -5,7 +5,7 @@ mkdir -p ./docs/img
 cp /app/mkdocs.yml .
 cp /app/favicon.ico ./docs/img
 cp --recursive /app/hooks .
-cp README.md ./docs/index.md
-cp LICESNE ./docs/license.md
+
+[ -f ./docs/license.md ] || cp LICESNE ./docs/license.md
 
 mkdocs build --clean
